Alexander

+++ resources/sections/AgregarLibros.php
@@ -92,8 +92,11 @@ if ($_POST) {
             $nombreTempImagen = $_FILES["libro-imagen"]["tmp_name"];
             $nombreImagen = $_FILES["libro-imagen"]["name"];
 
-            //si una nueva imagen es pasada al editar el libro tenemos que borrar la anterior
-            file_exists($rutaImg . $nombreImagendb) ? unlink($rutaImg . $nombreImagendb) : $rutaImg . $nombreImagendb;
+            //si una nueva imagen es pasada al editar el libro tenemos que borrar la anterior, excepto si es la de default
+            if ($nombreImagendb !== "default.png") {
+                file_exists($rutaImg . $nombreImagendb) ? unlink($rutaImg . $nombreImagendb) : $rutaImg . $nombreImagendb;
+            }
+
 
             $gestor->setImagen(moverImagen($nombreTempImagen, $rutaImg, $nombreImagen));
         } else {
